generated/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
openpyxl>=3.1.0
merlinquantum>=0.2.2

